Library

Compose.Mirror — Method.

Mirror(θ, x, y)

Mirror line passing through point (x,y) at angle θ (in radians).

source

Compose.Mirror — Method.

Mirror(θ)

Mirror(θ)=Mirror(θ, 0.5w, 0.5h)

source

Compose.Rotation — Method.

Rotation(θ, x, y)

Rotate all forms in context around point (x,y) by angle θ in radians.

source

Compose.Rotation — Method.

Rotation(θ)

Rotation(θ)=Rotation(θ, 0.5w, 0.5h)

source

Base.fill — Method.

fill(cs::AbstractArray)

Arguments can be passed in arrays in order to perform multiple drawing operations at once.

source

Base.fill — Method.

fill(c)

Define a fill color, where c can be a Colorant or String.

source

Compose.arc — Function.

arc(x, y, r, θ1, θ2, sector)

Define an arc with its center at (x,y), radius of r, between θ1 and θ2. sector (optional) is true or false, true for a pie sector, false for an arc. Arcs are drawn clockwise from θ1 to θ2.

source

Compose.arc — Function.

arc(xs::AbstractVector, ys::AbstractVector, rs::AbstractVector, θ1s::AbstractVector, θ2s::AbstractVector, sectors::AbstractVector)

Arguments can be passed in arrays in order to perform multiple drawing operations.

source

Compose.arrow — Method.

arrow(values::AbstractArray)

Arguments can be passed in arrays in order to perform multiple drawing operations at once.

source

Compose.arrow — Method.

arrow(value::Bool)

`arrow()` is a property of arcs, lines and curves. The color of the arrowhead is the same as `stroke()`, but for svg the results will be browser-dependent.

source

Compose.arrow — Method.

arrow()

`arrow() = arrow(true)`

source

Compose.bitmap — Function.

bitmap(mimes::AbstractArray, datas::AbstractArray, x0s::AbstractArray, y0s::AbstractArray, widths::AbstractArray, heights::AbstractArray)

Arguments can be passed in arrays in order to perform multiple drawing operations.

source

Compose.bitmap — Function.

bitmap(mime, data, x0, y0, width, height)

Define a bitmap of size widthxheight with its top left corner at the point (x, y).

source

Compose.circle — Function.

circle(xs::AbstractArray, ys::AbstractArray, rs::AbstractArray)

Arguments can be passed in arrays in order to perform multiple drawing operations.

source

Compose.circle — Function.

circle(x, y, r)

Define a circle with its center at (x,y) and a radius of r.

source

Compose.circle — Method.

circle()

Define a circle in the center of the current context with a diameter equal to the width of the context.

source

Compose.curve — Function.

curve(anchor0s::AbstractArray, ctrl0s::AbstractArray, ctrl1s::AbstractArray, anchor1s::AbstractArray)

Arguments can be passed in arrays in order to perform multiple drawing operations.

source

Compose.curve — Function.

curve(anchor0, ctrl0, ctrl1, anchor1)

Define a bezier curve between anchor0 and anchor1 with control points ctrl0 and ctrl1.

source

Compose.ellipse — Function.

ellipse(xs::AbstractArray, ys::AbstractArray, x_radiuses::AbstractArray, y_radiuses::AbstractArray)

Arguments can be passed in arrays in order to perform multiple drawing operations.

source

Compose.ellipse — Function.

ellipse(x, y, x_radius, y_radius)

Define an ellipse with its center at (x,y) with radii x_radius and y_radius.

source

Compose.ellipse — Method.

ellipse()

Define an ellipse in the center of the current context with x_radius=0.5w and y_radius=0.5h.

source

Compose.fillopacity — Method.

fillopacity(values::AbstractArray)

Arguments can be passed in arrays in order to perform multiple drawing operations at once.

source

Compose.fillopacity — Method.

fillopacity(value)

Define a fill opacity, where 0≤value≤1. For svg, nested contexts will inherit from parent contexts e.g. (context(), fillopacity(a), (context(), fill(c::String), circle())).

source

Compose.line — Function.

line(point_arrays::AbstractArray)

Arguments can be passed in arrays in order to perform multiple drawing operations at once.

source

Compose.line — Method.

line(points)

Define a line. points is an array of (x,y) tuples.

source

Compose.ngon — Function.

ngon(x, y, r, n::Int)

Define a n-sided polygon with its center at (x,y), and radius of r. For an upside-down ngon, use -r.

source

Compose.ngon — Function.

ngon(xs::AbstractVector, ys::AbstractVector, rs::AbstractVector, ns::AbstractVector{Int})

Arguments can be passed in arrays in order to perform multiple drawing operations at once.

source

Compose.polygon — Function.

polygon(point_arrays::AbstractArray)

Arguments can be passed in arrays in order to perform multiple drawing operations at once.

source

Compose.polygon — Method.

polygon(points)

Define a polygon. points is an array of (x,y) tuples that specify the corners of the polygon.

source

Compose.rectangle — Function.

rectangle(x0s::AbstractArray, y0s::AbstractArray, widths::AbstractArray, heights::AbstractArray)

Arguments can be passed in arrays in order to perform multiple drawing operations at once.

source

Compose.rectangle — Function.

rectangle(x0, y0, width, height)

Define a rectangle of size widthxheight with its top left corner at the point (x, y).

source

Compose.rectangle — Method.

rectangle()

Define a rectangle that fills the current context completely.

source

Compose.sector — Method.

sector(xs::AbstractVector, ys::AbstractVector, rs::AbstractVector, θ1s::AbstractVector, θ2s::AbstractVector)

Arguments can be passed in arrays in order to perform multiple drawing operations.

source

Compose.sector — Method.

sector(x, y, r, θ1, θ2)

Define a pie sector with its center at (x,y), radius of r, between θ1 and θ2.

source

Compose.star — Function.

star(x, y, r, n::Int, ratio)

Define a n-pointed star with its center at (x,y), outer radius of r, and inner radius equal to r*ratio. For an upside-down star, use -r.

source

Compose.star — Function.

star(xs::AbstractVector, ys::AbstractVector, rs::AbstractVector, ns::AbstractVector{Int}, ratios::AbstractVector{Float64})

Arguments can be passed in arrays in order to perform multiple drawing operations at once.

source

Compose.text — Function.

text(x, y, value [,halign::HAlignment [,valign::VAlignment [,rot::Rotation]]])

Draw the text value at the position (x,y) relative to the current context.

The default alignment of the text is hleft vbottom. The vertical and horizontal alignment is specified by passing hleft, hcenter or hright and vtop, vcenter or vbottom as values for halgin and valgin respectively.

source

Compose.text — Function.

text(xs::AbstractArray, ys::AbstractArray, values::AbstractArray [,haligns::HAlignment [,valigns::VAlignment [,rots::Rotation]]])

Arguments can be passed in arrays in order to perform multiple drawing operations at once.

source

Compose.xgon — Function.

xgon(x, y, r, n::Int, ratio)

Define a cross with n arms with its center at (x,y), outer radius of r, and inner radius equal to r*ratio. For an upside-down xgon, use -r.

source

Compose.xgon — Function.

xgon(xs::AbstractVector, ys::AbstractVector, rs::AbstractVector, ns::AbstractVector{Int}, ratios::AbstractVector{Float64})

Arguments can be passed in arrays in order to perform multiple drawing operations at once.

source

Compose.FormBatch — Type.

A form batch is a vectorized form with n primitives transformed into a simpler representation: one primitive repositioned n times.

On certain backends this leads to more efficient drawing. For example, SVG can be shortened by using <def> and <use> tags, and raster graphics can render the form primitive to a back buffer and blit it into place for faster drawing.

Batching is an optimization transform that happens at draw time. There's currently no mechanism to manually batch. E.g. contexts cannot have FormBatch children.

source

Compose.batch — Method.

Attempt to batch a form. Return a Nothing singleton if the Form could not be batched, and FormBatch object if the original form can be replaced.

source

Compose.canbatch — Method.

Some backends can more efficiently draw forms by batching. If so, they shuld define a similar method that returns true.

source

Compose.count_unique_primitives — Method.

Count the number of unique primitives in a property, stopping when max_count is exceeded.

source

Compose.excise_vector_children! — Method.

Remove and return vector forms and vector properties from the Context.

source

Compose.filter_redundant_offsets! — Method.

Produce a new array of offsets in which near duplicate values have been removed.

source

Compose.optimize_batching — Method.

Attempt to transform a tree into an equivalent tree that can more easily be batched.

What this does is look for patterns in which a long vector form is accompanied by a large vector property that has a relatively small number of unique values. If there are n unique values, we can split it into n contexts, each with a shorter vector form and only scalar properties.

source